refactor: only accept a key for setParams
It doesn't make much sense for setParams to accept a name as you'd not want to set params for a random route.
So this commit changes the signature so that only a key can be passed.

It also adds a 'source' property to actions which the reducer can use if key wasn't provided.