fix: swipe between child and parent TabNavigator (#881)
Cancels Gesture when swiping past either end of routes, which allows parent PanGestureHandler to become the gesture handler.
Provides a context-based mechanism for child PanGestureHandlers to pass their refs up to the parent in order to use in simultaneousHandlers and waitFor props.
Motivation
Fixes #880

Test plan
Create a nested tab navigator. Verify that swipes on the child, swipes on the parent, and swipes between child and parent all behave as expected.