chore: remove ios and android folders (#11020)
This PR removes prebuild `ios` and `android` folders. By default,
react-navigation uses Expo Go for development and testing. Prebuild in
the react-navigation repo is only used for testing the flipper plugin
and this causes confusion (the project is prebuilt but we use Expo Go
either way) and additional maintenance.

We motivate this decision by having less code in the repo, less
maintenance, and less confusion about project setup.