fix: avoid re-rendering screen when nested navigator state changes (#11547)
Previously the route object was recreated if the nested navigator's state changed inside a screen. This change avoids the unnecessary re-render by caching the `route` object as the `route.state` property isn't exposed to the user.
 
BREAKING CHANGE: This breaks `getFocusedRouteNameFromRoute` inside components as it requires a re-render to get the new value. However, it still works in `options` callback which is its intended use case.

---------

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>