feat: preloading in routers
# Preloading screens

## Problem

We noticed that it is common to have some heavy screens, which will need
to be loaded and some point. The obvious idea is to pre-load the screen
so it can live already rendered under the current screen.

## Action plan

This project is roughly divided into parts. 

First, the logic should include changes inside the routers (included in
this PR). Second, those changes should be reflected inside navigators
(or they should throw errors if unimplemented, e.g., in the native
stack).

### Routers

The navigation state should hold, together with `routes`, also
`preloadedRoutes: NavigationRoute<ParamList, keyof ParamList>[]` or
`preloadedRoutesKeys: string[];`.

That is the shape of the `PRELOAD` action:

```tsx
{
  type: 'PRELOAD';
  payload: {
    name: string;
    params?: object;
  };
  source?: string;
  target?: string; 
} 
```

That list will indicate to the navigator that those routes should be
readily available to show at any point.

Roughly, routers are split into two categories:

- Tab and Drawer routers
- If the route is not loaded yet, we should simply add a key to the
`preloadedRoutesKeys` array.
- If the route was already loaded with a matching ID, it should also get
added to the `preloadedRoutesKeys` (consequently) and replace the route
in `routes`.
- If the route was already loaded (the route with the name), but with no
matching ID, it should also get added to the `preloadedRoutesKeys` and
we should change the route under the proper index in `routes`.
- On navigating to that route, it should get removed from
`preloadedRoutesKeys`.
    - If the route is already focused, we warn and do nothing.
- if the route (with a matching ID) was already preloaded, calling
`PRELOAD` again should not update the `preloadedRoutesKeys` and it
should get reflected by updating the params in the route.
- Stack router
- If the route is not loaded yet, we should simply add a new route to
the `preloadedRoutes` array.
- If the route was already loaded (the route with the name), but with no
matching ID, it should also get added to the `preloadedRoutes` and not
modify `routes`.
- If the route was already loaded with a matching ID, it should also get
added to the `preloadedRoutes` (to make it clear for the navigator to
keep it loaded) and update params.
- On navigating to that route, it should get removed from
`preloadedRoutes`.
    - If the route is already focused, we warn and do nothing.
- if the route (with a matching ID) was already preloaded, calling
`PRELOAD` again should update the `preloadedRoutes` (params of the
route) and it should get reflected by updating the params in the route
if already in `routes`.

On top of that, we should have `DISMISS_PRELOAD` action, which (in both
cases) removes the route from the array.

```tsx
{
  type: 'DISMISS_PRELOAD';
  payload: {
    name: string;
    params?: object; // To get ID
  };
  source?: string;
  target?: string;
}
```

This PR includes those changes enriched with tests.