Revert "feat: enable preloading at material-top-tabs (#11755)" (#11756)
This reverts commit 58691ac973b40a93f4433dba0639dc3f45b21250.

Please provide enough information so that others can review your pull
request.

**Motivation**

Explain the **motivation** for making this change. What existing problem
does the pull request solve?

If this pull request addresses an existing issue, link to the issue. If
an issue is not present, describe the issue here.

**Test plan**

Describe the **steps to test this change** so that a reviewer can verify
it. Provide screenshots or videos if the change affects UI.

The change must pass lint, typescript and tests.