feat: introduce missing props from native-stack v5 (#11803)
**Motivation**

For some time now, people are complaining that current native stack does
not have some props that are included in v5, but not in v6.
I've deeply investigated the `props.tsx` file and I can make a
conclusion some of the things that are missing include:
- no `hideKeyboardOnSwipe` `gestureResponseDistance` property,
- no props for managing the form sheet - `sheetX` properties,
- missing support for `onWillAppear` (is it intentional?) and
`onGestureCancel` events.

This PR fixes all those things by adding them to `props.tsx` and their
support in `NativeStackView.native.tsx` file.
One thing I also wanted to mention is that some of the events from the
event map in `types.tsx` like `dismiss`, `appear` or
`headerHeightChange` were not included in changes. That's because I
think these ones should not be emitted directly from `navigation#emit`
and/or they might lose the support in future react-native-screens
versions. For example, `appear` event (which is currently deprecated)
should not be used and you should replace it with `transitionEnd` event
instead.

**Test plan**

You can test these props by checking how they affect screens after
adjusting them in Native Stack example. Remember about setting
`presentation` to `formSheet` while using `sheetX` props.

Some of the examples I've already prepared:

<details>
<summary>Using GestureResponseDistance on NewsFeed</summary>

This will enable changing the area of the gesture with a rectangle
starting from the ~middle of the screen down to bottom.

`NativeStack.tsx:139`
```tsx
      <Stack.Screen
        name="NewsFeed"
        component={NewsFeedScreen}
        options={{
          title: 'Feed',
          fullScreenGestureEnabled: true,
          gestureResponseDistance: {
            start: 40,
            end: 353,
            top: 400,
            bottom: 852,
          },
        }}
      />
```
</details>

<details>
<summary>Using sheet props on NewsFeed</summary>

This will change the NewsFeed screen to the form sheet with rounded
corners and maximum height to the middle.

`NativeStack.tsx:139`
```tsx
      <Stack.Screen
        name="NewsFeed"
        component={NewsFeedScreen}
        options={{
          title: 'Feed',
          presentation: 'formSheet',
          sheetAllowedDetents: 'medium',
          sheetCornerRadius: 32,
        }}
      />
```
</details>